Put all runtime options into their own module.
