+++ index.js
@@ -5,8 +5,11 @@
 const app = require('./app');
 const debug = require('debug')('meanwifi:server');
 const http = require('http');
+const runtime = require('./runtime');
 
-const port = normalizePort(process.env.PORT || '3000');
+runtime.printOptions();
+
+const port = runtime.port;
 app.set('port', port);
 
 const server = http.createServer(app);
@@ -14,25 +17,6 @@ const server = http.createServer(app);
 server.listen(port);
 server.on('error', onError);
 server.on('listening', onListening);
-
-/**
- * Normalize a port into a number, string, or false.
- */
-function normalizePort(val) {
-  let port = parseInt(val, 10);
-
-  if (isNaN(port)) {
-    // named pipe
-    return val;
-  }
-
-  if (port >= 0) {
-    // port number
-    return port;
-  }
-
-  return false;
-}
 
 /**
  * Event listener for HTTP server "error" event.
